% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_question.R
\name{ask_question}
\alias{ask_question}
\title{Ask question in the command line}
\usage{
ask_question(title, answers = c("y", "N"))
}
\arguments{
\item{title}{the question.}

\item{answers}{the vector for answers to the question. You may define as many
as possible.}
}
\value{
The selected answer as a character vector. If no answer returned,
e.g. pressed on Enter immediately after the prompt, a \code{NULL} is returned
instead.
}
\description{
Ask question in the command line
}
\examples{
\dontrun{
ask_question("Do you like me?")
ask_question("How much do you like this?", seq(5))
}
}
