% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_.R
\name{c_}
\alias{c_}
\title{Override base R allowing to put commas at the end (trailing comma)}
\usage{
c_(...)
}
\arguments{
\item{...}{objects to be passed onto [base::c].}
}
\description{
Override base R allowing to put commas at the end (trailing comma)
}
\examples{
c_("trailing", "commas", "are", "cool", )
}
\references{
\url{https://gist.githubusercontent.com/flodel/5283216/raw/a781f7f320bd11e41a2c81affe5bec6f226dbf5e/ok.comma.R}
}
