% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{calculate_sample_means}
\alias{calculate_sample_means}
\title{Calculating sample means from a population}
\usage{
calculate_sample_means(x, n, size, replace = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{an atomic numeric vector.}

\item{n}{number of samples to be taken.}

\item{size}{sample size to choose.}

\item{replace}{should sampling done with replacement? default value is FALSE.}

\item{na.rm}{remove NAs when calculating arithmetic mean. default value is TRUE.}
}
\value{
This function takes \code{n} number of samples with a size of \code{size} and
returns a vector with a length of \code{n}.
}
\description{
According to the central limit theorem, the mean of the samples taken from a
population would be normally distributed regardless of the distribution of
the population.
}
\examples{
data <- runif(1e3)
sm <- calculate_sample_means(data, 1e2, 40) # now look at the hist(sm)
}
