% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catby}
\alias{catby}
\title{Cat bold and yellow text to console}
\usage{
catby(...)
}
\arguments{
\item{...}{a valid \R object.}
}
\description{
Yellow is a nice distinct color and sometimes I need it especially when I am in
the middle of a development.
}
