% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catcl}
\alias{catcl}
\title{Cat with ANSI styles}
\usage{
catcl(..., color = NULL, style = NULL)
}
\arguments{
\item{...}{arguments passed by \code{\link{cat}}.}

\item{color}{color name.}

\item{style}{style name.}
}
\description{
Cat with ANSI styles
}
\section{Available options for the parameters}{
\subsection{for the \sQuote{color} parameter}{
\Sexpr[stage=install,results=rd]{
mmy::list_en_end(names(mmy:::ANSI_STYLES[["color"]]), add.quotes = TRUE)
}
}
\subsection{for the \sQuote{style} parameter}{
\Sexpr[stage=install,results=rd]{
mmy::list_en_end(names(mmy:::ANSI_STYLES[["style"]]), add.quotes = TRUE)
}
}
}
