% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_text.R
\name{comp_text}
\alias{comp_text}
\title{Turn a string compatible for directories}
\usage{
comp_text(text)
}
\arguments{
\item{text}{\code{character}.}
}
\description{
Turn a string compatible for directories
}
\details{
Takes the input string and do the following steps: turning it into
lower case, remove all punctuation, add hyphen between characters splited
into separate pieces.
}
