% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_int}
\alias{diff_int}
\title{Lagged differences for the integer values}
\usage{
diff_int(x)
}
\arguments{
\item{x}{an integer vector.}
}
\value{
Number of integers which equals to the length of the \code{x} minus one.
}
\description{
Lagged differences for the integer values
}
\examples{
diff_int(as.integer(c(1, 5, 3)))
}
