% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pattern.R
\name{extract_pattern}
\alias{extract_pattern}
\title{Extract pattern from a text object}
\usage{
extract_pattern(object, pattern = ".")
}
\arguments{
\item{object}{character.}

\item{pattern}{character. a PCRE pattern.}
}
\value{
a data frame if there is any match, else returns a logical \code{FALSE}.
}
\description{
Extract pattern from a text object
}
\examples{
\dontrun{
extract_pattern(c("a", "b", "c"), "c")
}
}
