% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-consecutive.R
\name{group_consecutive}
\alias{group_consecutive}
\title{Group sequences}
\usage{
group_consecutive(x, .early_exit = TRUE)
}
\arguments{
\item{x}{a numeric vector with a length of one.}

\item{.early_exit}{exits function early if there're no consecutive numbers. 
This argument is mainly for benchmarking purposes. Default value is \code{TRUE}.}
}
\value{
A list with a number of groups. The vectors in the list elements are the vector
\strong{indices}, not values.
}
\description{
Could be either increasing or decreasing sequences.
The order in the vector matters. There's no ordering done during the computation.

The missing values aren't handled gracefully. Be careful if the input data has 
any missing values.
}
\examples{
x <- c(16, 17, 18, 19, 100, 99, 98)
group_consecutive(x)
## missing one in the middle:
y <- c(seq(1, 3), seq(5, 8))
group_consecutive(y)
}
