% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-seq.R
\name{group_sequence}
\alias{group_sequence}
\title{Group sequences}
\usage{
group_sequence(vector)
}
\arguments{
\item{vector}{a numeric vector with a length of one.}
}
\description{
Could be either increasing or decreasing sequences.
}
\examples{
x <- c(16, 17, 18, 19, 100, 99, 98)
grps <- group_sequence(x)
## get actual values from indices:
sapply(grps, function(g) x[g])
## create a data.frame from groups:
data.frame(
group = seq_along(grps), 
indices = I(grps), 
elements = I(sapply(grps, function(g) x[g])),
stringsAsFactors = FALSE
)
}
