% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headtail}
\alias{headtail}
\title{Display head and tail of an object together}
\usage{
headtail(x, n = 3L, ...)
}
\arguments{
\item{x}{an object.}

\item{n}{a single integer. number of rows to be displayed per data i.e.
\code{n = 2} means four rows in total.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
TODO on progress....
}
\examples{
\dontrun{
headtail(iris)
}
}
