% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\alias{ht.data.frame}
\alias{ht.default}
\alias{ht.list}
\title{Print head-tail}
\usage{
ht(x, ...)

\method{ht}{data.frame}(x, n = 5)

\method{ht}{default}(x, n = 5)

\method{ht}{list}(x, n = 5)
}
\arguments{
\item{x}{an \R object.}

\item{n}{an integer indicating the number of first and last rows to be
printed. The default value is \code{5}.}
}
\description{
Print head-tail
}
\details{
The objects containing class of \code{data.frame} will
call \code{C} implementation and the other implementations calls \R functions.

The call cannot be assigned to any variable, and the result will be
\code{NULL (empty)} if it is done it so.
}
\examples{
## data.frame:
ht(mtcars)
## vector (default):
ht(LETTERS, n = 10)
## list:
ht(as.list(letters))
}
