% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_on_ci_platform}
\alias{is_on_ci_platform}
\title{Is the \R session running on a CI (continious integration) platform?}
\usage{
is_on_ci_platform()
}
\description{
Is the \R session running on a CI (continious integration) platform?
}
