% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-terminal.R
\name{is_terminal}
\alias{is_terminal}
\title{Check if session is in an "interactive" terminal}
\usage{
is_terminal()
}
\description{
Check if session is in an "interactive" terminal
}
\references{
\url{https://github.com/r-lib/prettycode/blob/7275e2a9f972c8837e070cccaf5ef514643cf607/R/utils.R#L4#L11}
}
