% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_en_end.R
\name{list_en_end}
\alias{list_en_end}
\title{List character vector by comma except the last one ending with 'and' phrase}
\usage{
list_en_end(x, add.quotes = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{add.quotes}{wrap each element with single quotes.}
}
\description{
List character vector by comma except the last one ending with 'and' phrase
}
\examples{
players <- LETTERS[1:10]
list_en_end(players)

list_en_end(round(rnorm(5), 3), add.quotes = TRUE)
}
