% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinside.R
\name{list_env_objects}
\alias{list_env_objects}
\title{List environment objects}
\usage{
list_env_objects(env = globalenv(), show.dot = FALSE)
}
\arguments{
\item{env}{which environment to look at.}

\item{show.dot}{display objects starting with dot.}
}
\description{
List environment objects
}
\details{
Similar to what RStudio environment pane displays.
}
