% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lse.R
\name{list_global_objects}
\alias{list_global_objects}
\title{List global environment objects}
\usage{
list_global_objects(env = ".GlobalEnv")
}
\arguments{
\item{env}{which environment to look at.}
}
\description{
List global environment objects
}
\details{
Similar to what RStudio environment pane displays.
}
