% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{machine_readable_name}
\alias{machine_readable_name}
\title{Turn a string compatible for machines}
\usage{
machine_readable_name(x)
}
\arguments{
\item{x}{a \code{character} vector.}
}
\description{
Non-compatible file names may create some nuisance in operating systems. This call
makes the string name ASCII compatible. Sadly no UTF support (yet). During the
convertion, non-ASCII names will be omitted.
}
\details{
Takes the input string and do the following steps: 
\itemize{
\item lower case
\item remove all punctuation (except hyphen)
\item add hyphen between characters splited into separate pieces
}
}
\examples{
\dontrun{
machine_readable_name(rownames(mtcars))
}
}
