% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmy-package.R
\docType{package}
\name{mmy-package}
\alias{mmy}
\alias{mmy-package}
\title{mmy: Custom miscellaneous R functions}
\description{
Custom solutions.
}
\author{
\strong{Maintainer}: Metin Yazici \email{stradivariusboul+r@gmail.com}

}
\keyword{internal}
