% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsprintf.R
\name{nsprintf}
\alias{nsprintf}
\title{Named sprintf}
\usage{
nsprintf(fmt, ..., phs = c("\\\\{\\\\{", "\\\\}\\\\}"))
}
\arguments{
\item{fmt}{format string including placeholders between shapes.}

\item{...}{the expressions used in placeholders.}

\item{phs}{placeholder shapes.}
}
\description{
Named sprintf
}
\examples{
nsprintf(
  "My name is {{name}}. I am {{number}} years old.",
  number = "25",
  name = "Apple"
)
}
\references{
Inspired from here:
\url{https://stackoverflow.com/a/55423080}
}
\seealso{
\href{https://cran.r-project.org/package=glue}{glue} and 
\href{https://cran.r-project.org/package=whisker}{whisker} packages.
}
