% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinside.R
\name{ObjectChecks}
\alias{ObjectChecks}
\alias{object_types}
\alias{check_language_object_types}
\title{Inspect types of R objects}
\usage{
object_types(..., echo = FALSE)

check_language_object_types(..., echo = FALSE)
}
\arguments{
\item{...}{valid \R object(s).}

\item{echo}{should objects be printed into the console?}
}
\description{
Inspect types of R objects
}
\details{
\itemize{
\item \code{class} displays \R's built-in S3 class.
\item \code{typeof} type of an object.
\item \code{mode} storage mode of an object.
\item \code{storage.mode} \emph{used when calling functions written in another
language.}
\item \code{sexp.type} internal \code{SEXPTYPE} reprentation, see
\code{\link{sexp.type}}.
}

\code{\link{mode}} and \code{\link{storage.mode}} are mainly used for the types
that are compatible with \emph{S}.

Substituted object names are stored in the \code{attributes}. See \emph{Examples}.
}
\examples{
object_types(1, 5L)
mmy::object_types(quote(for (i in seq(5L)) i))
types <- object_types(`$`, 1L, `[[<-`)
attributes(types)

quo <- quote(x <- 2)
check_language_object_types(quo, quo[[1]])

}
\references{
\href{https://cran.r-project.org/doc/manuals/r-release/R-lang.html}{R Language 
Definition manual}
}
