% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFilePath.R
\name{outputFilePath}
\alias{outputFilePath}
\title{Output character file path as R's platform-agnostic file.path}
\usage{
outputFilePath(path, platform = NULL)
}
\arguments{
\item{path}{file path as character.}

\item{platform}{name of the platform. Specify if you have a path separated by
different file separators which doesn't belong to the current platform where \R
runs. Select one of \code{unix} or \code{win} alternatives. If the argument is
\code{NULL}, which is the default value, a separator, which is specific to the
current platform, will automatically be selected \emph{(from
\code{.Platform$file.sep})}.}
}
\description{
Output character file path as R's platform-agnostic file.path
}
\examples{
path_nix <- "/var/log/wifi.log"
outputFilePath(path_nix)
}
