% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{pasteOutputAddin}
\alias{pasteOutputAddin}
\title{Paste Output RStudio Addin}
\usage{
pasteOutputAddin()
}
\description{
Why this addin? :: Helps e.g. when writing unit tests.
Why wrap it in R6? :: It keeps around tidy!
}
\details{
TODO Paste output as comment (a new endpoint)
TODO Reformat/reindent the pasted code
}
