% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d.R
\name{plot_1d}
\alias{plot_1d}
\title{One-dimensional plot}
\usage{
plot_1d(
  x,
  highlight = NULL,
  title = NULL,
  subtitle = NULL,
  display.max.min = TRUE,
  labels = TRUE,
  labels.at = range(x),
  labels.round = TRUE,
  ablines.at = NULL,
  linecolor = "lightgray",
  points.color = "darkgrey",
  alpha.points = FALSE,
  highlight.color = "red",
  abline.color = highlight.color,
  points.pch.type = 21,
  box = TRUE,
  grid = FALSE
)
}
\arguments{
\item{x}{a numeric vector.}

\item{highlight}{highlight}

\item{title}{title}

\item{subtitle}{subtitle}

\item{display.max.min}{display.max.min}

\item{labels}{labels}

\item{labels.at}{labels.at}

\item{labels.round}{round the displayed min and max values to 2 after
decimals.}

\item{ablines.at}{ablines.at}

\item{linecolor}{linecolor}

\item{points.color}{points.color}

\item{alpha.points}{alpha.points}

\item{highlight.color}{highlight.color}

\item{abline.color}{abline.color}

\item{points.pch.type}{points.pch.type}

\item{box}{box}

\item{grid}{add gridlines.}
}
\description{
This type of graph provides a batteries-included solution to plot one
dimensional graphs, which is also production ready.
}
\details{
One-dimensional graphs are useful in the cases:
\itemize{
\item plotting regression residuals,
\item plotting clustered data (e.g. k-means)
\item various coefficients etc.
}
}
\examples{
## simple vector:
x <- c(-0.1, 0.34, 0.55, 0.95)
plot_1d(x)
## with highlighting:
unif <- runif(1e2)
cond <- unif > 0.3 & unif < 0.5
plot_1d(
  x = unif,
  highlight = cond,
  alpha.points = TRUE,
  labels.at = mean(unif[cond]),
  title = "Uniform distribution",
  subtitle = "Subset values and the subset mean represented in red",
  highlight.color = "orange",
  box = FALSE,
  grid = TRUE
)
## multiple highlights:
norm <- rnorm(1e2)
cond1 <- norm < -0.8 & norm > -1.2
cond2 <- norm > 0.7 & norm < 1.0
cond1.mean <- mean(norm[cond1])
cond2.mean <- mean(norm[cond2])
plot_1d(
  x = norm,
  highlight = list(cond1, cond2),
  alpha.points = TRUE,
  labels.at = list(cond1.mean, cond2.mean),
  ablines.at = list(cond1.mean, cond2.mean),
  highlight.color = list("steelblue", "red"),
  title = "Normal distribution",
  subtitle = "Peak mean values in groups",
  box = FALSE,
  grid = TRUE
)
}
