% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_minmax.R
\name{rescale_minmax}
\alias{rescale_minmax}
\title{Feature scaling: Min-max normalization}
\usage{
rescale_minmax(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Feature scaling: Min-max normalization
}
\details{
The formula is as follows:
\deqn{
X_{scale} = \frac{X_i - X_{min}}{X_{max}-X_{min}}
}
}
\examples{
## Before rescaling:
plot(mtcars$mpg, mtcars$disp)
## After rescaling:
plot(rescale_minmax(mtcars$mpg), rescale_minmax(mtcars$disp))

}
