% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{row_names}
\alias{row_names}
\alias{un_row_names}
\alias{std_row_names}
\title{Dealing with rownames}
\usage{
un_row_names(x)

std_row_names(x)
}
\arguments{
\item{x}{a \code{data.frame}.}
}
\description{
Dealing with rownames
}
\details{
\enumerate{
\item \code{un_row_names} removes the row names of a data.frame.
\item \code{std_row_names} standardizes row names. In other words, it removes them
and attaches as a new column into a new \code{data.frame}.
}
}
\examples{
\dontrun{
un_row_names(mtcars)
}
}
