% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{rownames}
\alias{rownames}
\alias{un_rownames}
\alias{std_rownames}
\title{Dealing with rownames}
\usage{
un_rownames(x)

std_rownames(x, rowname = NULL)
}
\arguments{
\item{x}{data.frame.}

\item{rowname}{the column name for the rownames. Default value is \code{.rowname}.}

\item{x}{data.frame.}
}
\description{
Dealing with rownames
}
\details{
\enumerate{
\item \code{un_row_names} removes the row names of a data.frame.
\item \code{std_row_names} standardizes row names. In other words, it removes them
and attaches as a new column into a new \code{data.frame}.
}
}
\examples{
un_rownames(mtcars)
std_rownames(mtcars)
}
