% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sln.R
\name{search_list_names}
\alias{search_list_names}
\title{Search a keyword in a list names - fast?}
\usage{
search_list_names(list, query)
}
\arguments{
\item{list}{list object.}

\item{query}{keyword to look in the list names.}
}
\description{
Useful when working with large lists. Especially for the content return as
JSON type converted into R list type.
}
