% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinside.R
\name{see_object_types}
\alias{see_object_types}
\title{See types of R objects}
\usage{
see_object_types(...)
}
\arguments{
\item{...}{valid \R object(s).}
}
\description{
See types of R objects
}
\examples{
see_object_types(1, "a")
see_object_types(1, 5L)
see_object_types(as.name("mean"))
see_object_types(`(`)
see_object_types(`$`, 1L, `[[<-`)
}
