% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinside.R
\name{sexp.type}
\alias{sexp.type}
\title{Return \code{SEXPTYPE} from C internals}
\usage{
sexp.type(x)
}
\arguments{
\item{x}{any valid \R object.}
}
\description{
Return \code{SEXPTYPE} from C internals
}
\seealso{
\code{\link{pryr}{sexp_type}} returns a "similar output" as this call does.
However, \code{pryr} package does a lot more than showing underlying SEXP types.
}
