% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shinyprintf}
\alias{shinyprintf}
\title{Shiny "printf" style tracing for debugging purposes}
\usage{
shinyprintf(x, text = NULL, con = c("stdout", "stderr"))
}
\arguments{
\item{x}{variable.}

\item{text}{character. (\emph{optional}) description added to the variable
name. if empty, the default value is \code{NULL} therefore only the variable
name will be used.}

\item{con}{character. pipe connection. default is \code{stdout}.}
}
\description{
Put it inside the \code{shiny::observe} in case of reactive values are run.
Recommended to add the following snippet as \strong{shinyprinf}:
\code{observe(mmy::shinyprintf(${1:var}, text = NULL))}.
}
\references{
\url{https://shiny.rstudio.com/articles/debugging.html}
}
