% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{std_error}
\alias{std_error}
\title{Standard error of the mean}
\usage{
std_error(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
In the cases that we do not know about the population parameter of variation
(which is usually the case), we can take a number of samples from the population
and calculate each of these sample means that will always deviate from normal.
}
\details{
The standard error of the mean formula is as follows.

If the population standard deviation is known:

\deqn{SE_{\overline{x}} = \frac{\sigma}{\sqrt{n}}}

where
\eqn{\sigma}: standard deviation of the population
\eqn{n}: number of observations of the sample

If the population standard deviation is not known, the standard error of the mean
is \emph{approximately} equal to the \dQuote{sample standard deviation} divided by
the number of observations.
}
