% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{text_trunc}
\alias{text_trunc}
\title{Truncate longer texts}
\usage{
text_trunc(x, value = 15, symbol = "...", sep = "")
}
\arguments{
\item{x}{a vector.}

\item{value}{threshold for truncation.}

\item{symbol}{the suffix symbol for the truncation. Default: \code{...}}

\item{sep}{separator between the three-dots and truncated text.}
}
\description{
Truncate longer texts
}
\examples{
\dontrun{
text_trunc(c("apple", "this is a long text with more letters in it"))
text_trunc("The quick brown fox jumps over the lazy dog", 20, sep = " ")
}
}
