% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlist.R
\name{tlist}
\alias{tlist}
\title{An opinionated way to construct (tree-like) lists}
\usage{
\code{tlist()} \cr
\Sexpr[strip.white=FALSE,results=rd]{mmy:::tlist_example_tree()}
}
\description{
(WORK IN PROGRESS)
}
\details{
\itemize{
\item There're no list names in \code{tlist}, only indices.
\item The list is flat. The nodes staying at the same level are not wrapped in
another list, rather a new node has been created.
\item 'tlist's are designed to be used in small, not long stuctures. The semantics
doesn't allow to go super deep down.
}


As seen in the graph above,

\itemize{
\item The indexes surrounded by parentheses appearing in each line indicate
\emph{levels}, e.g. \code{(1)}.
\item The indexes placed in the bar appearing under the letters indicate
\emph{nodes}, e.g. \code{[5]}.
}
}
\examples{
## Instantiate a tlist to keep fruit data:
fruits <- tlist()
fruits$append("apple", 1, 1)
## Append a child
fruits$append("seed", 1, 3)
## Append siblings:
fruits$append(c("orange", "watermelon"), 1)
## Get the main list:
fruits$main()

}
