% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn-system-name-comp.R
\name{turn_system_name_comp}
\alias{turn_system_name_comp}
\title{Turn a string compatible for system directories}
\usage{
turn_system_name_comp(text)
}
\arguments{
\item{text}{\code{character}.}
}
\description{
Non-compatible file names may create some nuisance in operating systems. This call
makes the string name ASCII compatible. Sadly no UTF support (yet). During the
convertion, non-ASCII names will be omitted.
}
\details{
Takes the input string and do the following steps: 
\itemize{
\item lower case
\item remove all punctuation (except hyphen)
\item add hyphen between characters splited into separate pieces
}
}
