% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTesting.R
\name{UnitTesting}
\alias{UnitTesting}
\alias{test_suite}
\alias{is_error}
\alias{is_identical}
\alias{is_equal}
\title{A minimal unit testing suite}
\usage{
test_suite(section, ...)

is_error(call)

is_identical(actual, expected)

is_equal(actual, expected)
}
\description{
A minimal unit testing suite
}
