% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view.data.frame}
\alias{view.data.frame}
\title{View a data frame in the browser with reactable}
\usage{
\method{view}{data.frame}(x, ...)
}
\arguments{
\item{x}{a.data.frame or a matrix.}
}
\description{
Note: Pagination is must for performance for the large tables.
}
\references{
\url{https://glin.github.io/reactable/articles/examples.html}
}
