% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view.list}
\alias{view.list}
\title{View a list in the browser}
\usage{
\method{view}{list}(x, ...)
}
\description{
View a list in the browser
}
\references{
\url{https://cran.r-project.org/package=listviewer}
}
