% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-tbl.R
\name{view_tbl}
\alias{view_tbl}
\title{View a data frame in the browser with reactable}
\usage{
view_tbl(x)
}
\arguments{
\item{x}{a.data.frame or a matrix.}
}
\description{
Note: Pagination is must for performance for the large tables.
}
\references{
\url{https://glin.github.io/reactable/articles/examples.html}
}
