% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_debug_makevars_flags}
\alias{warn_debug_makevars_flags}
\title{A primitive way to see if Makevars has debug flags}
\usage{
warn_debug_makevars_flags()
}
\description{
A primitive way to see if Makevars has debug flags
}
\details{
The most common \emph{Makevars} flags for \R:
\itemize{
\item \code{CFLAGS}
\item \code{CXXFLAGS}
\item \code{CXX11FLAGS}
\item \code{FFLAGS}
\item \code{FCFLAGS}
}
}
