% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn-recommended-packages.R
\name{warn_recommended_packages}
\alias{warn_recommended_packages}
\title{Warn recommended packages}
\usage{
warn_recommended_packages(packages)
}
\arguments{
\item{packages}{a data.frame containing the metadata of the packages.}
}
\description{
Warn recommended packages
}
