% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_environment}
\alias{get_environment}
\title{Get environment variables inside a call}
\usage{
get_environment(which = 1L, exclude = NA_character_)
}
\arguments{
\item{which}{integer. the frame number in the call stack.}

\item{exclude}{character. which arguments should be excluded? It is possible
to exclude some variables by name.}
}
\description{
Returns all variables in a call environment. Made it possible with
\code{sys.frames()}, which accesses to the environments in the call stack.
}
