% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Head and tail of a data.frame}
\usage{
ht(x, n = 5)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{n}{an integer indicating the number of first and last rows to be
printed. The default value is \code{5}.}
}
\description{
Head and tail of a data.frame
}
\details{
Why creating such function? - I want to often glance on first and last rows
of a \code{data.frame} in R in a quick way.

The call cannot be assigned to any variable, and the result will be
\code{NULL (empty)} if it is done it so.

Why using \emph{R's C API} for such a simple function? - I saw this as an
opportunity to start learning R internals.

Your implementation isn't the most effective way of doing it. - I don't know
about this and it may be correct, but this's not a brain surgery either.
}
