% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\alias{ht.data.frame}
\alias{ht.default}
\alias{ht.list}
\title{Glance first and last rows}
\usage{
ht(x, n = 5)

\method{ht}{data.frame}(x, n = 5)

\method{ht}{default}(x, n = 5)

\method{ht}{list}(x, n = 5)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{n}{an integer indicating the number of first and last rows to be
printed. The default value is \code{5}.}
}
\description{
Glance first and last rows
}
\details{
The objects containing class of \code{data.frame} will
call C implementation and the other implementations calls R functions.

The call cannot be assigned to any variable, and the result will be
\code{NULL (empty)} if it is done it so.
}
