% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{pasteOutputAddin}
\alias{pasteOutputAddin}
\title{Paste Output RStudio Addin}
\usage{
pasteOutputAddin()
}
\description{
Paste Output RStudio Addin
}
\details{
If you are using Nvim-R, \code{:Rinsert dput(head(iris))} will work
for you. See here: \url{https://www.vim.org/scripts/script.php?script_id=2628}

Why this addin? :: Helps e.g. when writing unit tests.
Why wrap it in R6? :: It keeps around tidy!

TODO Paste output as unit test with expect_equal (a new endpoint)
TODO Reformat/reindent the pasted code
}
