% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rownames.R
\name{row_names}
\alias{row_names}
\alias{un_row_names}
\alias{stnd_row_names}
\title{Dealing with rownames}
\usage{
un_row_names(x)

stnd_row_names(x)
}
\arguments{
\item{x}{\code{data.frame}.}
}
\description{
Dealing with rownames
}
\details{
\itemize{
\item \code{un_row_names} removes the row names of a data.frame.
\item \code{stnd_row_names} standardizes row names. In other words, remove them and
attach as a new column in the \code{new data.frame}.
}
}
\examples{
\dontrun{
un_row_names(mtcars)
}
}
