% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_trunc.R
\name{text_trunc}
\alias{text_trunc}
\title{Truncate longer texts}
\usage{
text_trunc(x, level = 15, sep = "")
}
\arguments{
\item{x}{a vector.}

\item{level}{threshold for truncation.}

\item{sep}{separator between the three-dots and truncated text.}
}
\description{
Truncate longer texts
}
\examples{
\dontrun{
text_trunc(c("apple", "this is a long text with more letters in it"))
text_trunc("The quick brown fox jumps over the lazy dog", level = 20, sep = "\%")
}
}
