% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlist.R
\name{tlist}
\alias{tlist}
\title{An opinionated way to construct (tree-like) lists}
\usage{
tlist()
}
\description{
An opinionated way to construct (tree-like) lists
}
\details{
\itemize{
\item There're no list names in \code{tlist}, only indices.
\item The list is flat. The nodes staying at the same level are not wrapped in
another list, rather a new node has been created.
\item 'tlist's are designed to be used in small, not long stuctures. The semantics
doesn't allow to go super deep down.
}

________________________________________________________\cr
(1)...........A..........B..............C...............\cr
_____________[1]________[2]____________[3]______________\cr
............/.|.\........|......../...|...|...\.........\cr
(2)........A..B..C.......A.......A....B...C....D........\cr
_____________[1]________[2]____________[3]______________\cr
........../.\....|..../.|.|.\........|....|.../.|.\.....\cr
(3)......A...B...A...A..B.C..D.......A....A..A..B..C....\cr
_________[1]____[2]_____[3]_________[4]__[5]___[6]______\cr
(...)\cr

As seen in the graph above,

\itemize{
\item The indexes surrounded by parentheses appearing in each line indicate
\emph{levels}, e.g. \code{(1)}.
\item The indexes placed in the bar appearing under the letters indicate
\emph{nodes}, e.g. \code{[5]}.
}
}
\examples{
## Instantiate a tlist to keep fruit data:
fruits <- tlist()
fruits$append("apple", 1, 1)
## Append a child
fruits$append("seed", 1, 3)
## Append siblings:
fruits$append(c("orange", "watermelon"), 1)
## Get the main list:
fruits$main()

}
